Fixed transaction bugs

+++ SignalRStocksBackend/Hubs/StockHub.cs
@@ -4,7 +4,7 @@ using SignalRStocksBackend.Entities;
 
 namespace SignalRStocksBackend.Hubs
 {
-    public class StockHub: Hub
+    public class StockHub : Hub
     {
         private StockContext db;
         public StockHub(StockContext db)
@@ -19,6 +19,7 @@ namespace SignalRStocksBackend.Hubs
             if (transaction.IsUserBuy)
             {
                 var user = db.Users.Find(x => x.Name == transaction.Username);
+                if (user.Cash - (transaction.Amount * transaction.Price) < 0) throw new Exception("Price is too high");
                 user.Cash = user.Cash - (transaction.Amount * transaction.Price);
                 var userShare = db.UserShares.FirstOrDefault(x =>
                     x.Share.Name == transaction.ShareName && x.User.Name == transaction.Username);
@@ -36,22 +37,22 @@ namespace SignalRStocksBackend.Hubs
                 {
                     userShare.Amount = userShare.Amount + transaction.Amount;
                 }
-                
+
                 db.SaveChanges();
+
                 Clients.All.SendAsync("buy", transaction);
             }
             else
             {
                 var user = db.Users.Find(x => x.Name == transaction.Username);
-                user.Cash = user.Cash + (transaction.Amount * transaction.Price);
+
                 var userShare = db.UserShares.Find(x =>
                     x.User.Name == transaction.Username && x.Share.Name == transaction.ShareName);
-                
+
                 userShare.Amount = userShare.Amount - transaction.Amount;
-                if (userShare.Amount == 0)
-                {
-                    db.UserShares.Remove(userShare);
-                }
+                if (userShare.Amount == 0) db.UserShares.Remove(userShare);
+                if (userShare.Amount < 0) throw new Exception("Invalid Amount of Shares");
+                user.Cash = user.Cash + (transaction.Amount * transaction.Price);
                 Clients.All.SendAsync("sell", transaction);
                 db.SaveChanges();
             }
@@ -69,15 +70,6 @@ namespace SignalRStocksBackend.Hubs
             Clients.All.SendAsync("loggedOut", name);
         }
 
-        public override Task OnDisconnectedAsync(Exception? exception)
-        {
-            --usercounter;
-            var name = new NameDto();
-            name.Name = "unknown";
-            name.UserCounter = usercounter;
-            Clients.All.SendAsync("loggedOut", name);
 
-            return base.OnDisconnectedAsync(exception);
-        }
     }
 }
